Make Shared the default, add Isolated as backup.
